requirements.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-phone-verify==2.0.0
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
mysqlclient==1.4.6
nexmo==2.4.0
oauthlib==3.1.0
phonenumbers==8.12.2
PyJWT==1.7.1
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
twilio==6.39.0
urllib3==1.25.9
